Work around wrong the skills loading issue
Issue: the skills were all registered but one of them was the opsdroid
itself (no clue why yet). The skill list is then iterated and the
opsdroid object is a module and does not fit with further functions
expectations (that it is iterable).

Quick Fix: work around was to comment for now the part where the modules
is loaded, however it is also not loading skills which are not objects
(Skill classes). So one can only do classes skills.